+++ anubis-management-api/src/main.py
@@ -1,8 +1,11 @@
 from fastapi import Depends, FastAPI
 from tenants import routers as t
+from tenants import models as t_models
 from policies import routers as p
+from policies import models as p_models
 from version import ANUBIS_VERSION
 from fastapi.openapi.utils import get_openapi
+from database import engine
 
 from fastapi.middleware.cors import CORSMiddleware
 
@@ -33,6 +36,11 @@ app.add_middleware(
 app.include_router(t.router)
 app.include_router(p.router)
 
+t_models.Base.metadata.create_all(bind=engine)
+p_models.Base.metadata.create_all(bind=engine)
+
+p_models.insert_initial_agent_type_values()
+p_models.insert_initial_mode_values()
 
 @app.get("/v1/")
 async def v1_root():