+++ anubis-management-api/src/policies/models.py
@@ -12,15 +12,15 @@ import default
 
 
 policy_to_mode = Table('policy_to_mode', Base.metadata,
-                       Column('policy_id', Integer, ForeignKey('policies.id')),
-                       Column('mode_iri', Integer, ForeignKey('modes.iri'))
+                       Column('policy_id', String, ForeignKey('policies.id')),
+                       Column('mode_iri', String, ForeignKey('modes.iri'))
                        )
 
 
 policy_to_agent = Table(
     'policy_to_agent', Base.metadata, Column(
-        'policy_id', Integer, ForeignKey('policies.id')), Column(
-            'agent_iri', Integer, ForeignKey('agents.iri')))
+        'policy_id', String, ForeignKey('policies.id')), Column(
+            'agent_iri', String, ForeignKey('agents.iri')))
 
 
 class Policy(Base):
@@ -32,7 +32,7 @@ class Policy(Base):
     resource_type = Column(String, index=True)
     mode = relationship("Mode", secondary=policy_to_mode)
     service_path_id = Column(
-        Integer,
+        String,
         ForeignKey(
             'service_paths.id',
             ondelete="CASCADE"))
@@ -42,7 +42,7 @@ class Agent(Base):
     __tablename__ = "agents"
 
     iri = Column(String, primary_key=True, index=True)
-    type = Column(Integer, index=True)
+    type = Column(String, index=True)
 
 
 class Mode(Base):
@@ -59,19 +59,13 @@ class AgentType(Base):
     name = Column(String, index=True, unique=True)
 
 
-@event.listens_for(AgentType.__table__, 'after_create')
-def insert_initial_agent_type_values(*args, **kwargs):
+def insert_initial_agent_type_values():
     db = SessionLocal()
+    if len(db.query(AgentType).all()) > 0:
+        return
     db.add(AgentType(name='agent', iri=default.AGENT_IRI))
     db.add(AgentType(name='group', iri=default.AGENT_GROUP_IRI))
     db.add(AgentType(name='class', iri=default.AGENT_CLASS_IRI))
-    db.commit()
-    db.close()
-
-
-@event.listens_for(Agent.__table__, 'after_create')
-def insert_initial_agent_type_values(*args, **kwargs):
-    db = SessionLocal()
     db.add(
         Agent(
             iri=default.AUTHENTICATED_AGENT_ID,
@@ -84,9 +78,10 @@ def insert_initial_agent_type_values(*args, **kwargs):
     db.close()
 
 
-@event.listens_for(Mode.__table__, 'after_create')
-def insert_initial_mode_values(*args, **kwargs):
+def insert_initial_mode_values():
     db = SessionLocal()
+    if len(db.query(Mode).all()) > 0:
+        return
     db.add(Mode(iri=default.READ_MODE_IRI, name='read'))
     db.add(Mode(iri=default.WRITE_MODE_IRI, name='write'))
     db.add(Mode(iri=default.CONTROL_MODE_IRI, name='control'))